refactor: clean-up and split modals
